fix confirmation price

+++ app/Http/Controllers/API/TavsirController.php
@@ -404,15 +404,19 @@ class TavsirController extends Controller
     function orderConfirm(TsOrderConfirmRequest $request, $id)
     {
         $data = TransOrder::findOrfail($id);
-        $data->detil->whereNotIn('id', $request->detil_id)->each(function ($item) {
+        if($data->status != TransOrder::WAITING_CONFIRMATION)
+        {
+            return response()->json(['error' => 'Order '.$data->status], 500);
+        }
+        $dataConfirm = $data->detil->whereNotIn('id', $request->detil_id)->each(function ($item) {
             $item->delete();
         });
-        $sum = $data->detil->sum('total_price');
-        $data->sub_total = $sum;
+
+        $data->sub_total = $dataConfirm->sum('total_price');
         $data->total = $data->sub_total + $data->fee + $data->service_fee;
         $data->status = TransOrder::WAITING_PAYMENT;
         $data->save();
-
+        
         $data = TransOrder::findOrfail($id);
         return response()->json(new TsOrderResource($data));
     }
